==> code/Web_Scraping/1_wikipedia_ufc_data.ipynb <==
# **Scraping UFC Fight Outcomes and List of Events From Wikipedia**

import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import urllib.request as urllib2

### **Scrape List of All UFC Cards From Wikipedia**

# pull list of past ufc events from wikipedia
url = "https://en.wikipedia.org/wiki/List_of_UFC_events"
page = requests.get(url)
soup = BeautifulSoup(page.text, 'lxml')
rows = soup.find('table', id="Past_events").find_all('tr')
columns = [x.text.strip() for x in rows[0].find_all('th')]
past_events = pd.DataFrame([[x.text.strip() for x in row.find_all('td')] for row in rows[1:]], columns = columns)
urls = []
for row in rows[1:]:
    s = row.find_all('td')
    if  len(s) >= 1 and getattr(s[1], 'a', None) != None:
       urls.append('https://en.wikipedia.org' + s[1].a['href'].strip())
    else:
        urls.append(None)
past_events['url'] = urls
past_events = past_events[past_events['Attendance'] != 'Cancelled']
past_events.sort_values('#', ascending = False, inplace = True)
past_events['Date'] = pd.to_datetime(past_events['Date'])
del past_events['Ref.']
past_events.to_csv("../../data/wikipedia_data/wikipedia_all_ufc_events.csv", index = False)

past_events.head()

### **Scrape List of All UFC Fight Outcomes For Each Card from Wikipedia**

dfs = []
for card_index, url in enumerate(past_events['url']):
    if url != None:
        page = requests.get(url)
        soup = BeautifulSoup(page.text, 'lxml')
        
        if url == 'https://en.wikipedia.org/wiki/UFC_on_Fox:_Henderson_vs._Melendez':
            table = soup.find_all('table', class_ = 'toccolours')[2].tbody.find_all('tr')
        elif url == 'https://en.wikipedia.org/wiki/UFC_on_FX:_Belfort_vs._Bisping':
            table = soup.find_all('table', class_ = 'toccolours')[1].tbody.find_all('tr')
        elif url == 'https://en.wikipedia.org/wiki/UFC_on_Fox:_Evans_vs._Davis':
            table = soup.find_all('table', class_ = 'toccolours')[1].tbody.find_all('tr')    
        elif url == 'https://en.wikipedia.org/wiki/UFC_on_Fuel_TV:_Sanchez_vs._Ellenberger':
            table = soup.find_all('table', class_ = 'toccolours')[2].tbody.find_all('tr')   
        elif url == 'https://en.wikipedia.org/wiki/2012_in_UFC#UFC_on_Fuel_TV:_The_Korean_Zombie_vs._Poirier':
            table = soup.find_all('table', class_ = 'toccolours')[3].tbody.find_all('tr') 
        elif url == 'https://en.wikipedia.org/wiki/UFC_on_FX:_Johnson_vs._McCall':
            table = soup.find_all('table', class_ = 'toccolours')[4].tbody.find_all('tr')     
        elif past_events['#'].iloc[card_index] == '212':
            table = soup.find_all('table', class_ = 'toccolours')[5].tbody.find_all('tr')    
        else:
            table = soup.find('span', id = 'Results').parent.find_next_sibling().find('tbody').find_all('tr')
            
        card_type = table[0].find('th').text.strip()
        allrows = []
        for row in table[2:]:
            card_type_holder = row.find_all('th')
            if len(card_type_holder) == 1:
                card_type =  card_type_holder[0].text.strip()
                continue
            elif len(card_type_holder) > 0 and card_type_holder[0].text.strip().lower() == 'weight class':
                continue
            else:
                row_info = row.find_all('td')
                s = [x.text.strip() for x in row_info]
                if s[-1] != '':
                    ref = row_info[-1].a['href'].replace("#",'')
                    note = soup.find('li', id = ref).find('span', class_ = "reference-text").text
                    s[-1] = note
                s.append(card_type)
                try:
                    winner_link = "https://en.wikipedia.org" + row_info[1].a['href']
                except:
                    winner_link = ''
                try:
                    loser_link = "https://en.wikipedia.org" + row_info[3].a['href']
                except:
                    loser_link = ''
                s.append(winner_link)
                s.append(loser_link)
                allrows.append(s)
                

        df = pd.DataFrame(allrows, columns = ['WeightClass', 'Winner', 'Outcome', 'Loser', 'Method', 'Round', 'Time', 'Notes', 'Card', 'Winner_url', 'Loser_url'])
        df['wiki_url'] =url
        df['event_order'] = 1 + np.arange(df.shape[0]) 
        dfs.append(df)

## Clean Up The Data

all_fights = pd.concat(dfs)
print(all_fights.Outcome.value_counts())
all_fights.Outcome = all_fights.Outcome.str.replace('.', '')
print(all_fights.Outcome.value_counts())

sub = all_fights.Method.str.contains(pat = '(?i)sub') 
knockout = all_fights.Method.str.contains(pat = '(?i)ko') 
unam_decision = all_fights.Method.str.contains(pat = '(?i)dec') & all_fights.Method.str.contains(pat = '(?i)(unanimous|majority)')
split_decision = all_fights.Method.str.contains(pat = '(?i)dec') & all_fights.Method.str.contains(pat = '(?i)split')
outcomes = pd.concat([sub, knockout, unam_decision, split_decision], axis=1)
display(all_fights.loc[outcomes.apply("sum", axis = 1) > 1, :])
sub[knockout] = False
outcomes = pd.concat([sub, knockout, unam_decision, split_decision], axis=1)
print(sum(outcomes.apply("sum", axis = 1) > 1))

all_fights.loc[sub, 'Method_Cleaned'] = "Sub"
all_fights.loc[knockout, 'Method_Cleaned'] = "KO"
all_fights.loc[unam_decision, 'Method_Cleaned'] = "Unanimous Decision"
all_fights.loc[split_decision, 'Method_Cleaned'] = "Split Decision"
print(all_fights.Method_Cleaned.value_counts(dropna= False))
display(all_fights.loc[(all_fights.Outcome == 'def') & (all_fights.Method_Cleaned.isna()), :])


print(all_fights.Card.value_counts(dropna= False))
main = all_fights.Card.str.contains('(?i)main')
prelim = all_fights.Card.str.contains('(?i)prelim')
print(sum(main & prelim))
all_fights.loc[main, 'Card_Cleaned'] = "Main Event"
all_fights.loc[prelim, 'Card_Cleaned'] = "Prelim"
print(all_fights.Card_Cleaned.value_counts(dropna= False))

# Standardize fighter names and determine if champion or interim champion is fighting

cleaner_winner = all_fights['Winner'].map(lambda x: x.split('(', 1))
cleaner_winner = cleaner_winner.map(lambda x: [x[0].strip(), x[1].strip().lower()] if len(x) > 1 else [x[0].strip()])
clean_winner = cleaner_winner.map(lambda x: x[0])

cleaner_loser = all_fights['Loser'].map(lambda x: x.split('(', 1))
cleaner_loser= cleaner_loser.map(lambda x: [x[0].strip(), x[1].strip().lower()] if len(x) > 1 else [x[0].strip()])
clean_loser = cleaner_loser.map(lambda x: x[0])

champ = cleaner_winner.map(lambda x: x[0] if len(x) > 1 and x[1] == 'c)' else '')
champ = np.where(champ == '', cleaner_loser.map(lambda x: x[0] if len(x) > 1 and x[1] == 'c)' else ''), champ)
interim_champ = cleaner_winner.map(lambda x: x[0] if len(x) > 1 and x[1] == 'ic)' else '')
interim_champ = np.where(interim_champ  == '', cleaner_loser.map(lambda x: x[0] if len(x) > 1 and x[1] == 'ic)' else ''), interim_champ)


all_fights['Winner_Cleaned'] = clean_winner
all_fights['Loser_Cleaned'] = clean_loser
all_fights['Champion'] = champ
all_fights['Interim_Champion'] = interim_champ
display(all_fights.head())

all_fights.to_csv("../../data/wikipedia_data/wikipedia_all_ufc_fight_outcomes.csv", index = False)
