==> code/Web_Scraping/2_bestfightodds_final_ufc_betting_odds.ipynb <==
# **Scraping UFC Betting Data From BestFightOdds**

Data consists of the final quoted odds on UFC fights for each betting site listed on bestfightodds.com.
All bet types are included

import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import urllib.parse
import re

### **Get BestFightOdds URL for Each UFC Event in Wikipedia**

events = pd.read_csv('../../data/wikipedia_data/wikipedia_all_ufc_events.csv')
search = events['Event'].apply(lambda x: 'https://www.bestfightodds.com/search?query=' + urllib.parse.quote(x))
events['Date'] = pd.to_datetime(events['Date'])
fightodds_url = [''] * events.shape[0]

for i in range(events.shape[0]):  
    date_of_card = events.iloc[i]['Date']
    page = requests.get(search.iloc[i])
    soup = BeautifulSoup(page.text, 'lxml')
    
    # skip if no possible matches found
    if soup.find('table') == None or page.text.lower().find("No betting lines available for this event".lower()) != -1:
        fightodds_url[i] = ''
        continue 
    
    # if query redirects to exact match take query as url
    if soup.find('tr', class_ = 'even')  != None:
        fightodds_url[i] = search.iloc[i]
        continue
        
    table = soup.find_all('table', class_ = 'content-list') [-1].find_all('tr')
    final_url = ''
    for p in table:
        if pd.to_datetime(p.find('td').text) == date_of_card:
            final_url = 'https://www.bestfightodds.com'+ p.find_all('td')[1].a['href']
            fightodds_url[i] = final_url
            break
    
    # in some cases time zones cause dates to be off by a day. 
    if final_url == '':
        for p in table:
            if np.abs((pd.to_datetime(p.find('td').text) - date_of_card) / pd.to_timedelta(1, unit='D')) <= 1 :
                final_url = 'https://www.bestfightodds.com'+ p.find_all('td')[1].a['href']
                fightodds_url[i] = final_url
                break
            
events['fight_odds_url'] =  fightodds_url

events.head()

# get coverage of how many ufc events in Wikipedia have bestfightodds urls
events['fight_odds_url'].groupby(events['Date'].apply(lambda x: x.year)).apply(lambda x: (x != '').sum())/events['fight_odds_url'].groupby(events['Date'].apply(lambda x: x.year)).size()

# view missing events
pd.set_option('display.max_rows', 500)
events[(events['fight_odds_url'] == '') & ( events['Date'].apply(lambda x: x.year) >= 2008) ]

# fill in missing events manually
events.loc[events['Event'] == 'UFC Fight Night: Błachowicz vs. Jacaré', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-on-espn-22-blachowicz-vs-jacare-1755'
events.loc[events['Event'] == 'UFC Fight Night: Błachowicz vs. Santos', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-on-espn-3-blachowicz-vs-santos-1638'
events.loc[events['Event'] == 'UFC on Fox: Evans vs. Davis', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-on-fox-2-482'
events.loc[events['Event'] == 'UFC Fight Night: Shields vs. Ellenberger', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-fight-night-25-battle-on-the-bayou-406'
events.loc[events['Event'] == 'UFC Live: Kongo vs. Barry', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-on-versus-4-379'
events.loc[events['Event'] == 'UFC Live: Jones vs. Matyushenko', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-on-versus-2-281'
events.loc[events['Event'] == 'The Ultimate Fighter: Team McGregor vs. Team Faber Finale', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-the-ultimate-fighter-22-finale-edgar-vs-mendes-1011'
events.loc[events['Event'] == 'The Ultimate Fighter: Heavy Hitters Finale', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-the-ultimate-fighter-28-finale-dos-anjos-vs-usman-1586'
events.loc[events['Event'] == 'The Ultimate Fighter: American Top Team vs. Blackzilians Finale', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-the-ultimate-fighter-21-finale-ellenberger-vs-thompson-971'
events.loc[events['Event'] == 'The Ultimate Fighter: Undefeated Finale', 'fight_odds_url'] = 'https://www.bestfightodds.com/events/ufc-the-ultimate-fighter-27-finale-tavares-vs-adesanya-1503'

# get coverage of % ufc events are available
print(events['fight_odds_url'].groupby(events['Date'].apply(lambda x: x.year)).apply(lambda x: (x != '').sum())/events['fight_odds_url'].groupby(events['Date'].apply(lambda x: x.year)).size())
print(sum((events.Date  >= pd.to_datetime("Jan 1, 2008")) & (events.fight_odds_url.isna() | (events.fight_odds_url == ''))))
events = events[~(events.fight_odds_url.isna() | (events.fight_odds_url == ''))]
events.to_csv("../../data/bestfightodds_data/bestfightodds_urls.csv", index = False)


# As a quality check for later pull all closing odds 

i.e. to make sure the final odds in the time series match the closing odds

# pull list of past ufc moneylines from bestfightodds
dfs = []
events = events.loc[(events['fight_odds_url'] != '') & 
                    (events['fight_odds_url'].notna()), :]
for card_index, url in enumerate(events['fight_odds_url']):
    page = requests.get(url)
    soup = BeautifulSoup(page.text, 'lxml')
    columns = ['class']+[t.text for t in soup.find_all('table',class_ = 'odds-table')[1].thead.tr.find_all('th')][0:-1]
    rowtable = soup.find_all('table',class_ = 'odds-table')[1].tbody.find_all('tr')
    rows = [([row['class'][0]] + [row.find('th').text] \
             + [t.text.replace('▲','').replace('▼','') for t in row.find_all('td')])[0:-2] for row in rowtable]
    df = pd.DataFrame(rows, columns = columns)
    df.rename(columns={'':'Bet'}, inplace=True)
    fight1 = []
    fight2 = []
    fighter1 = ''
    fighter2 = ''
    
    hit_event_props = False
    for i in range(df.shape[0]):
        
        if df.iloc[i]['Bet'].lower() == 'event props':
            fighter1 = 'Event Prop'
            fighter2 = fighter1
            hit_event_props = True
            
        elif hit_event_props == False and df.iloc[i]['class'] == 'even':
            fighter1 = df.iloc[i]['Bet']
            fighter2 = df.iloc[i+1]['Bet']
            
        fight1.append(fighter1)
        fight2.append(fighter2)
        
    df.insert(0,'fighter1', fight1)
    df.insert(1,'fighter2', fight2)
    df.insert(2,'bet_type', np.where((df['class'] == 'even') | (df['class'] == 'odd'),'Ordinary','Prop'))
    df.loc[df['fighter1'] == 'Event Prop','bet_type'] = 'Event Prop'
    del df['class']
    # get rid of row showing event props if it's present
    if hit_event_props == True:
        df.drop(df[(df['Bet'].map(lambda x: x.lower()) == 'event props')].index, inplace=True)
    df['url'] = url
    df['Card_Date'] = events['Date'].iloc[card_index]
    df['Event'] = events['Event'].iloc[card_index]
    dfs.append(df)

all_odds = pd.concat(dfs)
all_odds.rename({'William\xa0H.':'William_H', 'SportsInt.':'SportsInt'}, inplace = True, axis = 1)
betting_sites= ['5Dimes','BetDSI','BookMaker','SportBet','Bet365','Bovada', 
                'Sportsbook','William_H','Pinnacle','SportsInt','BetOnline','Intertops']
for col in betting_sites:
    all_odds[col] = pd.to_numeric(all_odds[col], errors = 'coerce')
all_odds.to_csv("../../data/bestfightodds_data/moneyline_data_at_close.csv", index = False)
